everything works

+++ controllers/contacts.js
@@ -1,5 +1,6 @@
 import asyncHandler from "express-async-handler";
 import Contact from "../model/Contact.js";
+import User from "../model/User.js";
 
 //create
 
@@ -12,14 +13,14 @@ export const createContact = asyncHandler(async (req, res)=>{
         throw new Error("All fields are mandatory")
     };
 
-    const contacts = await Contact.create({name,email,contact});
+    const contacts = await Contact.create({name,email,contact,user_id: req.user.id,});
     res.status(201).json(contacts);
 
 
 })
 //get
 export const getContacts = asyncHandler(async (req, res)=>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 
 })
@@ -30,6 +31,11 @@ export const updateContact = asyncHandler(async (req, res)=>{
     if (!contact) {
         res.status(400);
         throw new Error("Contact not found");        
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Unauthorised, you can't update other peoples contact!!");
+        
     }
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new:true});
 
@@ -53,6 +59,11 @@ export const deleteContact = asyncHandler (async (req, res) =>{
     if (!contact) {
         res.status(400);
         throw new Error("Contact not found");        
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Unauthorised, you can't delete other peoples contact!!");
+        
     }
     const deleteContact = await Contact.findOneAndDelete(req.params.id);
     res.status(200).json({message: `Delete contact for ${req.params.id}`})
